basic gui onclick event

+++ AGame/zSystem.cpp
@@ -18,7 +18,7 @@ void SystemDatabase::SystemDatabaseUpdate(const float& dt) {
 				// loop through
 				for (auto& chunk : archetype.second->_chunk_database) {
 					system.second->_current_chunk = chunk.get();
-					for (int i = 0; i < chunk->_number_of_entities; ++i) {
+					for (int i = 0; i < (int)chunk->_number_of_entities; ++i) {
 						if (system.second->_current_chunk->_active_flags[i]) {
 							system.second->_current_id = i;
 							system.second->UpdateComponent();
